Se modifica el tipo del atributo porcentaje, de integer a float

+++ database/migrations/2021_09_21_050609_create_socios_table.php
@@ -17,7 +17,7 @@ class CreateSociosTable extends Migration
             $table->id();
             $table->string('apellido');
             $table->string('nombre');
-            $table->integer('porcentaje');
+            $table->float('porcentaje');
             $table->timestamps();
 
             $table->foreign('id_sociedad')
